fix: Change the start index of the current index to 0
Since the previous attempt of implementation start the index from 1, it confuse  me for future usage of current index. So I decided to revert it back to 0, but the index that is displayed in the prev and next display property start from 1. I do for the user can track the what index display they currently in (since the display counter is start from 1 not 0).